backend/requirements.txt
absl-py==1.4.0
aniso8601==8.0.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
black==23.1.0
blis==0.7.9
bokeh==3.0.3
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
click==7.1.2
colorama==0.4.6
comm==0.1.2
confection==0.0.3
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
Cython==0.29.33
debugpy==1.6.6
decorator==5.1.1
dill==0.3.6
distro==1.8.0
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.8.0
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
idna==3.4
ipykernel==6.21.2
ipython==8.10.0
itsdangerous==1.1.0
jedi==0.18.2
Jinja2==2.11.2
jupyter_client==8.0.3
jupyter_core==5.2.0
keras==2.11.0
kiwisolver==1.4.4
langcodes==3.3.0
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==1.1.1
matplotlib==3.7.0
matplotlib-inline==0.1.6
murmurhash==1.0.9
mypy-extensions==1.0.0
mysql-connector-python==8.0.32
nest-asyncio==1.5.6
nlp==0.4.0
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.1.1
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.1
parso==0.8.3
pathspec==0.11.1
pathy==0.6.2
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
preshed==3.0.8
prompt-toolkit==3.0.37
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pygame==2.1.2
Pygments==2.14.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2020.1
pywin32==305
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.3
six==1.15.0
smart-open==5.2.1
spacy==3.4.2
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.3.18
srsly==2.4.5
stack-data==0.6.2
tabula-py==2.7.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
tesseract==0.1.3
thinc==8.1.5
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typer==0.4.2
typing_extensions==4.4.0
urllib3==1.26.12
wasabi==0.10.1
wcwidth==0.2.6
Werkzeug==1.0.1
wrapt==1.14.1
xxhash==3.1.0
xyzservices==2023.2.0